fix(types): Do not apply `scale` twice on primitive
Up to now, `scale` is applied in `create_primitive()` to the
primitive after its creation. The reason for this was a bug in
Blender which ignored the `scale` argument in
`bpy.ops.mesh.primitive_..._add()`. This seems to be fixed by now,
so the scaling inside BlenderProc results in applying the scaling
twice.

This commit fixes this bug in BlenderProc by deleting the second
scaling.