feat(object): Improve `sample_pose()`
`sample_pose()` now allows to define its behavior when the current
object could not be placed without collision. Additionally it
returns a `Dict` containing the objects to sample as keys and a
Tuple with the number of attempts to place the respective object
and whether it was succesful or not.

Fixes #435