fix(blenderproc): Fix __init__.py for Windows
Currently, the stack summary is check on startup of BlenderProc,
in order to determine whether it has been called correctly (i.e.
either via the `blenderproc` command or the command line script.
This has not worked on Windows, since the stack trace differs
from the one on Linux, and hence an exception is thrown by
__init__.py.

This commit fixes this issue by performing different checks of
the stack trace depending on the current OS. Additionally, the
check for Linux was rewritten to be understandable more easily.

Fixes #358